RR works for longer inputs, sqrtf to sqrt
but the sqrt still limits the input size

+++ AccRQA-GPU-RR-ER.cu
@@ -51,7 +51,7 @@ __global__ void GPU_RQA_RR_kernel(
 	extern __shared__ int s_local_RR[]; //local recurrent rate
 	unsigned long long int pos_x, pos_y;
 	__shared__ int s_sums[NTHREADS];
-	int local_RR[10]; // WIP: Set this equivalent to emb
+	unsigned long long int local_RR[10]; // WIP: Set this equivalent to emb
 
 	// Set local_RR array to zero, otherwise the RR count accumulates for different thresholds.
 	for(int i_t = 0; i_t < emb; i_t++) {
@@ -64,7 +64,7 @@ __global__ void GPU_RQA_RR_kernel(
 	int stride = blockDim.x * gridDim.x; // To loop ove the threads one grid size at a time.
 	while (i < size*(size-1)/2) {
 		// Calculate x,y position in the upper triangle of the RQA matrix from the linear index, i
-		pos_x = (unsigned long long int) size - 2 - (int) (sqrtf(-8.0 * i + 4.0 * size * (size - 1) - 7.0) / 2.0 - 0.5);
+		pos_x = (unsigned long long int) size - 2 - (unsigned long long int) (sqrt(-8.0 * i + 4.0 * size * (size - 1) - 7.0) / 2.0 - 0.5);
 		pos_y = (unsigned long long int) i + pos_x + 1 - size * (size - 1) / 2 + (size - pos_x) * ((size - pos_x) - 1) / 2;
 		
 		for (int k = 0; k < emb; ++k)
